Replace few MSVC-only functions with std and fio

+++ src/libs/sound_service/src/sound_service.cpp
@@ -851,13 +851,13 @@ void SoundService::SetActiveWithFade(const bool active)
 
     FMOD::ChannelGroup* mastergroup;
     CHECKFMODERR(system->getMasterChannelGroup(&mastergroup));
-    uint64_t parentclock;
+    unsigned long long parentclock; // type should match with getDSPClock param from fmod.hpp
     int rate;
     system->getSoftwareFormat(&rate, nullptr, nullptr);
     mastergroup->getDSPClock(nullptr, &parentclock);
 
     const auto dsp_clock_start = parentclock;
-    const auto dsp_clock_end = static_cast<uint64_t>(parentclock + fadeTimeInSeconds * rate);
+    const auto dsp_clock_end = static_cast<unsigned long long>(parentclock + fadeTimeInSeconds * rate);
 
     if (active)
     {
